Move variation helper functions to shared dir
